REFACTORIZANDO BACKEND
