+++ BackEnd/src/main/resources/application.properties
@@ -6,7 +6,7 @@ spring.datasource.password=${DB_PASSWORD}
 
 hibernate.dialect=org.hibernate.dialect.HSQLDialect
 
-#spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=update
 
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
